fix(ui): spring cleaning before merging to develop
