+++ aether-ui/aether/ui/context_processors.py
@@ -7,7 +7,6 @@ def ui_context(request):
     context = {
         'dev_mode': settings.DEBUG,
         'app_name': settings.APP_NAME,
-        'kernel_url': settings.AETHER_APPS['kernel']['url'],
     }
 
     return context
