--- aether-ui/aether/ui/templates/pages/tokens.html
@@ -1,41 +1,0 @@
-{% extends 'base.html' %}
-{% load i18n %}
-{% load ui_tags %}
-
-
-{% block title %}{{ app_name }} - {% trans 'User tokens' %}{% endblock title %}
-
-
-{% block content %}
-  <div class="page-content p-5">
-    <div class="page-title mb-4">
-      {% trans 'App tokens for' %}
-      <span class="username ml-2">
-        <b>{{ request.user|get_fullname }}</b>
-      </span>
-    </div>
-
-    <table class="table table-striped table-hover">
-      <thead>
-        <tr>
-          <th>{% trans 'Server' %}</th>
-          <th>{% trans 'Token' %}</th>
-          <th>{% trans 'URL' %}</th>
-        </tr>
-      </thead>
-
-      <tbody>
-        <tr>
-          <td>Aether Kernel</td>
-          <td>{{ request.user|get_token:'kernel' }}</td>
-          <td>
-            <span class="url-name">
-              <a href="{{ kernel_url }}" target="_blank">{{ kernel_url }}</a>
-            </span>
-          </td>
-        </tr>
-      </tbody>
-    </table>
-
-  </div>
-{% endblock content %}
