+++ aether-ui/aether/ui/tests/test_context_processors.py
@@ -11,5 +11,4 @@ class ContextProcessorsTests(TestCase):
         self.assertEqual(ui_context(request), {
             'dev_mode': False,
             'app_name': 'aether',
-            'kernel_url': 'http://kernel-test:9000',
         })
