+++ aether-ui/aether/ui/tests/test_settings.py
@@ -16,7 +16,3 @@ class SettingsTest(TestCase):
         self.assertEqual(settings.ROOT_URLCONF, 'aether.ui.urls')
         self.assertEqual(settings.WSGI_APPLICATION, 'aether.ui.wsgi.application')
         self.assertEqual(settings.APP_NAME, 'aether')
-        self.assertEqual(settings.AETHER_MODULES, ['kernel'])
-
-        self.assertIn('kernel', settings.AETHER_APPS)
-        self.assertEqual(settings.AETHER_APPS['kernel']['url'], 'http://kernel-test:9000')
