+++ aether-ui/conf/extras/initial.json
@@ -12,12 +12,5 @@
       "is_staff": true,
       "is_active": true
     }
-  },
-  {
-    "model": "ui.usertokens",
-    "pk": 1,
-    "fields": {
-      "kernel_token": "a2d6bc20ad16ec8e715f2f42f54eb00cbbea2d24"
-    }
   }
 ]
