+++ aether-ui/entrypoint.sh
@@ -65,16 +65,6 @@ setup_initial_data() {
   ./manage.py loaddata ./conf/extras/initial.json
 }
 
-setup_projects() {
-  # Wait for kernel to become available
-  until $(curl --output /dev/null --silent --head $AETHER_KERNEL_URL); do
-    >&2 echo 'Waiting for Aether kernel...'
-    sleep 1
-  done
-  # Set up Aether and Ui projects and ensure that they are in sync
-  ./manage.py setup_aether_project
-}
-
 setup_prod() {
   # check if vars exist
   ./conf/check_vars.sh
@@ -175,7 +165,6 @@ case "$1" in
   start )
     setup_db
     setup_prod
-    setup_projects
 
     # remove previous files
     rm -r -f ${BUNDLES_DIR}
